Version 1.8 : il a vachement plu des commentaires, là !

+++ source/repos/Parce que c'est notre projet/Parce que c'est notre projet/Plateau.cs
@@ -124,69 +124,77 @@ namespace Parce_que_c_est_notre_projet
             return s;
         }
         /// <summary>
-        /// Test de la condition d'adjacence dans un plateau
+        /// Test de la condition d'adjacence et la présence d'un mot dans un plateau
         /// </summary>
         /// <param name="mot">Mot à analyser</param>
         /// <param name="posimot">Position du caractère du mot en cours d'étude (par défaut, 0)</param>
-        /// <param name="positionMot">Liste comportant les cases déjà étudies (par défaut nulle)</param>
-        /// <returns></returns>
+        /// <param name="positionMot">Liste comportant les cases déjà étudies (par défaut null)</param>
+        /// <returns>Booléen : true = le mot respecte la condition d'adjecence et est présent, false sinon</returns>
         public bool Test_Plateau(string mot, int posimot, List<int> positionMot = null)
         {
-            Console.WriteLine(mot);
             bool estPresent = false;
-            if(posimot == 0)
+            if(mot != null && mot.Length != 0) // On vérifie si le mot peut être analysé
             {
-                for(int i = 0; i < this.valeurSup.GetLength(0) && !estPresent; i++)
+                // Analyse pour la première position
+                // On recherche d'abord si la première lettre du mot est bien dans le plateau
+                if (posimot == 0) 
                 {
-                    for (int j = 0; j < this.valeurSup.GetLength(1) && !estPresent; j++)
+                    // Parcours de toutes les cases du plateau pour trouver la première lettre du mot
+                    for (int i = 0; i < this.valeurSup.GetLength(0) && !estPresent; i++)
                     {
-                        if(this.valeurSup[i,j] == Convert.ToString(mot[0]))
+                        for (int j = 0; j < this.valeurSup.GetLength(1) && !estPresent; j++)
                         {
-                            positionMot = new List<int>();
-                            positionMot.Add(i);
-                            positionMot.Add(j);
-                            estPresent = Test_Plateau(mot, posimot + 1, positionMot);
-                            if (!estPresent)
+                            if (this.valeurSup[i, j] == Convert.ToString(mot[0]))
                             {
-                                positionMot.RemoveRange(0, positionMot.Count);
-                            }
-                            else
-                            {
-                                return estPresent;
+                                // La première lettre du mot est dans le plateau, à la case [i, j]
+                                positionMot = new List<int>(); // On initialise la liste comportant les coordonnées des cases déjà étudiées
+                                // On ajoute à cette liste la position actuelle (d'abord i, puis j)
+                                positionMot.Add(i);
+                                positionMot.Add(j);
+                                estPresent = Test_Plateau(mot, posimot + 1, positionMot); // On va ensuite chercher les autres lettres du mot
+                                if (!estPresent) // Si la recherche de la ligne ci-dessus a échoué, on retire les positions de la liste
+                                {
+                                    positionMot.RemoveRange(0, positionMot.Count);
+                                }
+                                else
+                                {
+                                    // Si on a trouvé le mot, on retourne directement la valeur true
+                                    return estPresent;
+                                }
                             }
                         }
                     }
                 }
-            }
-            else if (posimot < mot.Length - 1)
-            {
-                for (int i = positionMot[positionMot.Count - 2] - 1; i <= positionMot[positionMot.Count - 2] + 1 && !estPresent; i++)
+                // On recherche ensuite toutes les autres lettres (dernière lettre exclue du else if)
+                else if (posimot < mot.Length - 1) 
                 {
-                    for (int j = positionMot[positionMot.Count - 1] - 1; j <= positionMot[positionMot.Count - 1] + 1 && !estPresent; j++)
+                    // Analyse des cases voisines à celle que l'on a trouvée précédemment (et qui nous est communiquée par la liste positionMot)
+                    for (int i = positionMot[positionMot.Count - 2] - 1; i <= positionMot[positionMot.Count - 2] + 1 && !estPresent; i++)
                     {
-                        if (((i >= 0 && i < this.valeurSup.GetLength(0)) && (j >= 0 && j < this.valeurSup.GetLength(1))) && (i != positionMot[positionMot.Count - 2] || j != positionMot[positionMot.Count - 1]))
+                        for (int j = positionMot[positionMot.Count - 1] - 1; j <= positionMot[positionMot.Count - 1] + 1 && !estPresent; j++)
                         {
-                            Console.WriteLine(posimot + " : " + i + ", " + j);
-                            if(this.valeurSup[i, j] == Convert.ToString(mot[posimot]))
+                            if (((i >= 0 && i < this.valeurSup.GetLength(0)) && (j >= 0 && j < this.valeurSup.GetLength(1))) && (i != positionMot[positionMot.Count - 2] || j != positionMot[positionMot.Count - 1]) && this.valeurSup[i, j] == Convert.ToString(mot[posimot]))
                             {
-                                Console.WriteLine("OK pour la position " + i + ", " + j);
+                                // On a trouvé une lettre qui est égale à la posimot-ième lettre de notre mot, et qui n'est pas sur la case sur laquelle on est centré
+                                // (On vérifie aussi si les coordonnées ne sont pas out of range)
                                 bool test = true;
-                                for (int k = 0; (k < positionMot.Count - 1 && test); k += 2)
+                                for (int k = 0; (k < positionMot.Count - 1 && test); k += 2) // On teste si la case correspondant aux conditions ci-dessus n'est pas déjà dans le chemin en cours d'étude (éviter les retours en arrière et doublons)
                                 {
                                     if (i == positionMot[k] && j == positionMot[k + 1])
                                     {
                                         test = false;
                                     }
                                 }
-                                if (test)
+                                if (test) // La lettre n'appartient pas à la liste des cases du chemin en cours, on peut donc aller travailler dessus
                                 {
+                                    // On ajoute à cette liste la position actuelle (d'abord i, puis j)
                                     positionMot.Add(i);
                                     positionMot.Add(j);
-                                    estPresent = Test_Plateau(mot, posimot + 1, positionMot);
-                                    positionMot.RemoveAt(positionMot.Count - 1);
+                                    estPresent = Test_Plateau(mot, posimot + 1, positionMot); // On va ensuite chercher les autres lettres du mot
+                                    positionMot.RemoveAt(positionMot.Count - 1); // Par défaut, on retire ces positions de la liste (pour éviter des tests inutiles)
                                     positionMot.RemoveAt(positionMot.Count - 1);
                                 }
-                                if (estPresent)
+                                if (estPresent) // Si on a en retour true, on le retourne immédiatement, mettant fin à la récursivité
                                 {
                                     return estPresent;
                                 }
@@ -194,35 +202,38 @@ namespace Parce_que_c_est_notre_projet
                         }
                     }
                 }
-            }
-            else
-            {
-                for (int i = positionMot[positionMot.Count - 2] - 1; i <= positionMot[positionMot.Count - 2] + 1 && !estPresent; i++)
+                // On recherche enfin la dernière lettre
+                else
                 {
-                    for (int j = positionMot[positionMot.Count - 1] - 1; j <= positionMot[positionMot.Count - 1] + 1 && !estPresent; j++)
+                    // Analyse des cases voisines à celle que l'on a trouvée précédemment (et qui nous est communiquée par la liste positionMot)
+                    for (int i = positionMot[positionMot.Count - 2] - 1; i <= positionMot[positionMot.Count - 2] + 1 && !estPresent; i++)
                     {
-                        if (((i >= 0 && i < this.valeurSup.GetLength(0)) && (j >= 0 && j < this.valeurSup.GetLength(1))) && (i != positionMot[positionMot.Count - 2] || j != positionMot[positionMot.Count - 1]))
+                        for (int j = positionMot[positionMot.Count - 1] - 1; j <= positionMot[positionMot.Count - 1] + 1 && !estPresent; j++)
                         {
-                            Console.WriteLine(posimot + " : " + i + ", " + j);
-                            bool test = true;
-                            for (int k = 0; k < positionMot.Count - 1; k += 2)
+                            if (((i >= 0 && i < this.valeurSup.GetLength(0)) && (j >= 0 && j < this.valeurSup.GetLength(1))) && (i != positionMot[positionMot.Count - 2] || j != positionMot[positionMot.Count - 1]) && this.valeurSup[i, j] == Convert.ToString(mot[posimot]))
                             {
-                                if (i == positionMot[k] && j == positionMot[k + 1])
+                                // On a trouvé une lettre qui est égale à la posimot-ième lettre de notre mot, et qui n'est pas sur la case sur laquelle on est centré
+                                // (On vérifie aussi si les coordonnées ne sont pas out of range)
+                                bool test = true;
+                                for (int k = 0; k < positionMot.Count - 1; k += 2) // On teste si la case correspondant aux conditions ci-dessus n'est pas déjà dans le chemin en cours d'étude (éviter les retours en arrière et doublons)
                                 {
-                                    test = false;
+                                    if (i == positionMot[k] && j == positionMot[k + 1])
+                                    {
+                                        test = false;
+                                    }
                                 }
-                            }
-                            if (test && this.valeurSup[i, j] == Convert.ToString(mot[posimot]))
-                            {
-                                Console.WriteLine("GO");
-                                return true;
-                                break;
+                                if (test) // Si la dernière lettre est dans le voisinage de l'avant dernière et n'a pas été parcourue, on termine la récursivité, en retournant true
+                                {
+                                    return true;
+                                    break;
+                                }
                             }
                         }
                     }
                 }
             }
-            return estPresent;
+            
+            return estPresent; // Dans le cas où l'on a pas trouvé le mot (ou il ne correspond pas aux critères de longueur), on retourne false par défaut
         }
     }
 }
